validate.py should be robust
and it should not fail with exception on invalid data,
it should rather report which lines are invalid.

This commit prevents exceptions when a CoNLL-U contains
e.g. just one column instead of 10 columns.
It is not very elegant, I admit.
Maybe it would be better to just jump to the next tree
after reporting less than 10 columns.