bugfix: UnboundLocalError: local variable 'head' referenced before assignment
I got this error when trying to validate original GUM files with --coref.
I've changed `GRP` to `eid`, in `global.Entity`.
The `global.Entity` header still does not include `head`,
so this is reported as [L6 Coref spurious-global-entity], which is OK.
However, then validate.py fails because accessing the `head` variable before assignment.